102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,25 +1,27 @@
 #include <stdio.h>
 
 /**
- * main - Prints the first 50 Fibonacci
+ * main - Prints the first 50 Fibonacci numbers separated by commas.
  * Return: Always 0.
  */
 int main(void)
 {
-	int n = 50;
-	int first = 1, second = 2, next;
-	int i;
+int n = 50;
+long int first = 1, second = 2, next;
+int i;
 
-	printf("%d", first);
+printf("%ld", first);
 
-	for (i = 1; i < n; i++)
-	{
-		printf(", %d", second);
-		next = first + second;
-		first = second;
-		second = next;
-	}
-	printf("\n");
-	return (0);
+for (i = 1; i < n; i++)
+{
+printf(", %ld", second);
+next = first + second;
+first = second;
+second = next;
+}
+
+printf("\n");
+
+return (0);
 }
